fix(ci): close history projection boundaries
