+++ .buildchain/kfd/kfd-1/release-gate.json
@@ -88,13 +88,13 @@
       "standard": "KFD-1",
       "standardKey": "kfd-1",
       "result": "passed",
-      "preBuildWitnessSha256": "42f4dc60de7f20eacc5bfa45f339e0d853134776b107c7ecda88f8b7ccd6e0e1",
+      "preBuildWitnessSha256": "18e63c39f5462ccbb670e41a789de1adb9fd5c79df1315e5aa4e69a12b92942d",
       "sourceHashes": {
-        "sha256": "3643ebcaae13ca2182f47821a473c5fd8f36fdd6497e392e5e580b38b9e3cad6",
+        "sha256": "f1d4bdb42f1903b713d715d269d0d626788334d454cd7c3e6ec18139cd659e4d",
         "surfaceCount": 21
       },
       "artifactHashes": {
-        "sha256": "c3e76c4fb03c37b2af379d396c867f3650abd6f4ea693d0127e385106436afb8",
+        "sha256": "5cc897bf7bc0525dac3f88db7415b8a99c2dcead50d2b79ff1073584b3f85f37",
         "surfaceCount": 21
       },
       "witness": {
@@ -134,7 +134,7 @@
         "sourceVerificationRequired": true,
         "canonicalPolicy": {
           "path": "framework/contract/kungfu-agent-first-canonical-policy.json",
-          "sha256": "641f7353036fb0e749fc8a3bf013798e73d879efb11e4fe2ab4f03cf1953ad6e"
+          "sha256": "81200b98aadeccf122a664f0e333089fcb097a2977036aa327119f5af3ba46f5"
         },
         "registry": {
           "path": "framework/contract/kungfu-contracts.registry.json",
@@ -304,9 +304,9 @@
           {
             "name": "kungfu-primitive-catalog",
             "sourcePath": "framework/primitive/kungfu-primitive-catalog.contract.json",
-            "sourceSha256": "a8844e4f8d5397e8fae766b13a7cc2f9c619477a90e16c272d24363ce6e51544",
+            "sourceSha256": "8435c627be1b8eb62fe7e714479eff5626e8a82418c6f8b656318986cb0d2b38",
             "artifactPath": "config/primitive/kungfu-primitive-catalog.contract.json",
-            "expectedSha256": "a8844e4f8d5397e8fae766b13a7cc2f9c619477a90e16c272d24363ce6e51544",
+            "expectedSha256": "8435c627be1b8eb62fe7e714479eff5626e8a82418c6f8b656318986cb0d2b38",
             "byteForByte": true
           }
         ],
@@ -549,8 +549,8 @@
           {
             "name": "kungfu-primitive-catalog",
             "sourcePath": "framework/primitive/kungfu-primitive-catalog.contract.json",
-            "expectedSha256": "a8844e4f8d5397e8fae766b13a7cc2f9c619477a90e16c272d24363ce6e51544",
-            "actualSha256": "a8844e4f8d5397e8fae766b13a7cc2f9c619477a90e16c272d24363ce6e51544",
+            "expectedSha256": "8435c627be1b8eb62fe7e714479eff5626e8a82418c6f8b656318986cb0d2b38",
+            "actualSha256": "8435c627be1b8eb62fe7e714479eff5626e8a82418c6f8b656318986cb0d2b38",
             "status": "passed",
             "reason": ""
           }
@@ -783,10 +783,10 @@
           {
             "name": "kungfu-primitive-catalog",
             "sourcePath": "framework/primitive/kungfu-primitive-catalog.contract.json",
-            "sourceSha256": "a8844e4f8d5397e8fae766b13a7cc2f9c619477a90e16c272d24363ce6e51544",
+            "sourceSha256": "8435c627be1b8eb62fe7e714479eff5626e8a82418c6f8b656318986cb0d2b38",
             "artifactPath": "config/primitive/kungfu-primitive-catalog.contract.json",
-            "expectedSha256": "a8844e4f8d5397e8fae766b13a7cc2f9c619477a90e16c272d24363ce6e51544",
-            "actualSha256": "a8844e4f8d5397e8fae766b13a7cc2f9c619477a90e16c272d24363ce6e51544",
+            "expectedSha256": "8435c627be1b8eb62fe7e714479eff5626e8a82418c6f8b656318986cb0d2b38",
+            "actualSha256": "8435c627be1b8eb62fe7e714479eff5626e8a82418c6f8b656318986cb0d2b38",
             "byteForByte": true,
             "status": "passed",
             "reason": ""
