+++ config/kungfu-agent-first-canonical-policy.json
@@ -523,13 +523,13 @@
       },
       "source": {
         "path": "framework/primitive/kungfu-primitive-catalog.contract.json",
-        "sha256": "sha256:a8844e4f8d5397e8fae766b13a7cc2f9c619477a90e16c272d24363ce6e51544",
-        "renderedSha256": "sha256:a8844e4f8d5397e8fae766b13a7cc2f9c619477a90e16c272d24363ce6e51544",
+        "sha256": "sha256:8435c627be1b8eb62fe7e714479eff5626e8a82418c6f8b656318986cb0d2b38",
+        "renderedSha256": "sha256:8435c627be1b8eb62fe7e714479eff5626e8a82418c6f8b656318986cb0d2b38",
         "byteForByte": true
       },
       "artifact": {
         "path": "config/primitive/kungfu-primitive-catalog.contract.json",
-        "expectedSha256": "sha256:a8844e4f8d5397e8fae766b13a7cc2f9c619477a90e16c272d24363ce6e51544"
+        "expectedSha256": "sha256:8435c627be1b8eb62fe7e714479eff5626e8a82418c6f8b656318986cb0d2b38"
       }
     }
   ]
