+++ docs/adr/KF-ADR-019fbe04-6548-7b3a-8358-22c8fe8238a9.md
@@ -5,7 +5,7 @@ adr_id: KF-ADR-019fbe04-6548-7b3a-8358-22c8fe8238a9
 decision_status: accepted
 implementation_status: staged
 implementation_prs: []
-qualification_refs: [framework/exit/kungfu-exit-bundle.contract.json, framework/core/src/python/kungfu/exit_bundle.py, framework/core/tests/python/test_exit_bundle.py, scripts/check-exit-history-surfaces.test.mjs, framework/tui/src/main.tsx, framework/gui/src/renderer/src/main.tsx]
+qualification_refs: [framework/exit/kungfu-exit-bundle.contract.json, framework/core/src/python/kungfu/exit_bundle.py, framework/core/src/python/kungfu/exit_verifier.py, framework/core/tests/python/test_exit_bundle.py, scripts/check-exit-history-surfaces.test.mjs, framework/tui/src/main.tsx, framework/gui/src/renderer/src/main.tsx, framework/gui/src/runtime-status.ts]
 review_state: self-reviewed
 sensitivity: public
 sources: [local-files, user-consensus]
