+++ docs/architecture/adr-map.md
@@ -273,7 +273,7 @@ review`; evidence is the number of declared qualification references.
 | KF · b9bd | [Kungfu owns native Work Control vocabulary](../adr/KF-ADR-019f9771-4c20-7e2c-8e7c-3f3cb3f1b9bd.md) | kungfu-native-work-control | accepted / implemented / self-reviewed | 21 | [KF · d8bc](../adr/KF-ADR-019f86da-4f90-7a66-b427-f4bcd638d8bc.md) (shared terms: kungfu)<br>[KF · d412](../adr/KF-ADR-019f86da-4f90-7f7b-90be-c002b024d412.md) (shared terms: control)<br>[KF · cb76](../adr/KF-ADR-019f86da-4f90-713d-8626-d70bca82cb76.md) (shared terms: kungfu)<br>[SHIFU · f011](../adr/SHIFU-ADR-019f86da-4f90-79a1-bc85-4b542fecf011.md) (shared terms: control) |
 | KF · 7ec0 | [Maintainability governance uses read-only composite regression ratchets](../adr/KF-ADR-019f9f04-f8bd-7002-a702-1b9f66827ec0.md) | maintainability-governance | accepted / implemented / self-reviewed | 31 | [KF · a40c](../adr/KF-ADR-019f8fb8-579b-78d5-9ebb-da03bb9aa40c.md) (shared terms: governance) |
 | KF · 90d2 | [First-party longitudinal evolution map](../adr/KF-ADR-019fa773-aae4-7f50-80a5-ce53b08490d2.md) | first-party-longitudinal-evolution-map | accepted / implemented / self-reviewed | 7 | [KF · 86f1](../adr/KF-ADR-019f86da-4f90-7e5e-ae22-2a8fc24086f1.md) (shared terms: first) |
-| KF · 38a9 | [Exit history surfaces are observers over unchanged domain authority](../adr/KF-ADR-019fbe04-6548-7b3a-8358-22c8fe8238a9.md) | durable-history-data-protection | accepted / staged / self-reviewed | 6 | [KF · d844](../adr/KF-ADR-019f86da-4f90-7008-b154-d488eb39d844.md) (shared terms: authority, domain, exit)<br>[SHIFU · 7ace](../adr/SHIFU-ADR-019f86ff-a8d6-7431-ae05-0ec95fdb7ace.md) (shared terms: authority)<br>[KF · 7733](../adr/KF-ADR-019f8822-1d7a-7594-adea-65ad12c47733.md) (shared terms: domain)<br>[KF · e7d5](../adr/KF-ADR-019f86da-4f90-76a1-8eda-6e49fa70e7d5.md) (shared terms: over) |
+| KF · 38a9 | [Exit history surfaces are observers over unchanged domain authority](../adr/KF-ADR-019fbe04-6548-7b3a-8358-22c8fe8238a9.md) | durable-history-data-protection | accepted / staged / self-reviewed | 8 | [KF · d844](../adr/KF-ADR-019f86da-4f90-7008-b154-d488eb39d844.md) (shared terms: authority, domain, exit)<br>[SHIFU · 7ace](../adr/SHIFU-ADR-019f86ff-a8d6-7431-ae05-0ec95fdb7ace.md) (shared terms: authority)<br>[KF · 7733](../adr/KF-ADR-019f8822-1d7a-7594-adea-65ad12c47733.md) (shared terms: domain)<br>[KF · e7d5](../adr/KF-ADR-019f86da-4f90-76a1-8eda-6e49fa70e7d5.md) (shared terms: over) |
 | SHIFU · c58b | [Repository root discovery and jurisdiction](../adr/SHIFU-ADR-019f86da-4f90-78f2-9256-43ef0fe3c58b.md) | shifu-repo-root-discovery | accepted / partial / unreviewed | 0 | [SHIFU · f011](../adr/SHIFU-ADR-019f86da-4f90-79a1-bc85-4b542fecf011.md) (shared terms: shifu)<br>[KF · 1491](../adr/KF-ADR-019f86da-4f90-765c-9723-069718911491.md) (shared terms: root)<br>[KF · abd2](../adr/KF-ADR-019f86da-4f90-7626-861e-3fdee887abd2.md) (shared terms: shifu) |
 | SHIFU · f011 | [Gate control plane contract](../adr/SHIFU-ADR-019f86da-4f90-79a1-bc85-4b542fecf011.md) | shifu-gate-control-plane | accepted / implemented / self-reviewed | 17 | [KF · d412](../adr/KF-ADR-019f86da-4f90-7f7b-90be-c002b024d412.md) (shared terms: control)<br>[KF · 8aec](../adr/KF-ADR-019f917f-d116-70e8-b4a1-2e0209598aec.md) (shared terms: plane)<br>[KF · 7733](../adr/KF-ADR-019f8822-1d7a-7594-adea-65ad12c47733.md) (shared terms: contract)<br>[KF · fe03](../adr/KF-ADR-019f86da-4f90-7d81-90a0-d144fc27fe03.md) (shared terms: contract) |
 | SHIFU · 7ace | [ADR identity is distributed UUIDv7 authority](../adr/SHIFU-ADR-019f86ff-a8d6-7431-ae05-0ec95fdb7ace.md) | — | accepted / implemented / maintainer-reviewed | 5 | [SHIFU · 35aa](../adr/SHIFU-ADR-019fab1a-2853-737e-8c67-a9b1aa9035aa.md) (shared terms: authority, identity)<br>[KF · d844](../adr/KF-ADR-019f86da-4f90-7008-b154-d488eb39d844.md) (shared terms: authority)<br>[KF · 38a9](../adr/KF-ADR-019fbe04-6548-7b3a-8358-22c8fe8238a9.md) (shared terms: authority)<br>[KF · a40c](../adr/KF-ADR-019f8fb8-579b-78d5-9ebb-da03bb9aa40c.md) (shared terms: authority) |
