+++ framework/gui/src/renderer/src/main.tsx
@@ -51,8 +51,11 @@ import {
   profileHomeId,
 } from '../../navigation';
 import {
+  EXIT_HISTORY_STATUS_FALLBACK,
+  type ExitHistoryStatus,
   type RuntimeStatusResult,
   deriveWorkspaceRuntimePresentation,
+  observeExitHistoryStatus,
 } from '../../runtime-status';
 import {
   GLOBAL_WORK_OBSERVER_EVENT_CHANNEL,
@@ -1105,6 +1108,14 @@ function App() {
   const startupViewApplied = React.useRef(false);
 
   React.useEffect(() => {
+    if (!runtime.ok) {
+      setHistoryStatus(EXIT_HISTORY_STATUS_FALLBACK);
+      return undefined;
+    }
+    return observeExitHistoryStatus(window, guiKungfuCliArgs, setHistoryStatus);
+  }, [runtime.ok]);
+
+  React.useEffect(() => {
     if (startupViewApplied.current || !config) return;
     startupViewApplied.current = true;
     const agent = config.config.agent as
@@ -1181,60 +1192,6 @@ function App() {
       active = false;
     };
   }, []);
-
-  React.useEffect(() => {
-    if (!runtime.ok) {
-      setHistoryStatus(EXIT_HISTORY_STATUS_FALLBACK);
-      return undefined;
-    }
-    type ExecFile = (
-      file: string,
-      args: string[],
-      options: {
-        encoding: 'utf8';
-        env: Record<string, string | undefined>;
-        maxBuffer: number;
-      },
-      callback: (error: Error | null, stdout: string) => void,
-    ) => void;
-    const { execFile } = window.require('node:child_process') as {
-      execFile: ExecFile;
-    };
-    const env: Record<string, string | undefined> = {
-      ...window.process.env,
-      KUNGFU_AS_VARIANT: undefined,
-    };
-    const bin =
-      env.KUNGFU_CLI_BIN ||
-      env.KUNGFU_BIN ||
-      (window.process.platform === 'win32' ? 'kungfu.exe' : 'kungfu');
-    let active = true;
-    execFile(
-      bin,
-      guiKungfuCliArgs(env, ['exit', 'history', 'status', '--json']),
-      { encoding: 'utf8', env, maxBuffer: 2 * 1024 * 1024 },
-      (error, stdout) => {
-        if (!active) return;
-        if (error) {
-          setHistoryStatus(EXIT_HISTORY_STATUS_FALLBACK);
-          return;
-        }
-        try {
-          const value = JSON.parse(stdout) as ExitHistoryStatus;
-          setHistoryStatus(
-            value.ok && value.nextActions?.length
-              ? value
-              : EXIT_HISTORY_STATUS_FALLBACK,
-          );
-        } catch {
-          setHistoryStatus(EXIT_HISTORY_STATUS_FALLBACK);
-        }
-      },
-    );
-    return () => {
-      active = false;
-    };
-  }, [runtime.ok]);
 
   React.useEffect(() => {
     if (
