+++ framework/primitive/kungfu-primitive-catalog.contract.json
@@ -27,7 +27,7 @@
     },
     "kfd1Contracts": {
       "path": "framework/contract/kungfu-contracts.registry.json",
-      "root": "sha256:1b51661a257538c04d4097132a2352b90e05645423ffb4f761deacd5a19016e5"
+      "root": "sha256:86079535834c818ba06972f3a1ff1bf8188c876746fd54d29ee04947f6de867b"
     },
     "invariants": {
       "path": "framework/invariant/kungfu-invariant.registry.json",
@@ -1697,5 +1697,5 @@
       "artifacts": []
     }
   ],
-  "catalogRoot": "sha256:11e16815b3a26a2ddc042bf3000046b687f9bc6898ec9c110fc803935f722509"
+  "catalogRoot": "sha256:38d704b093be58e1e9424e8191595c4ad78b673abbfb8fe5656236cb3be3dc76"
 }
