+++ framework/registry/contract.registry-envelope.json
@@ -17,7 +17,7 @@
       "id": "contract-registry",
       "source": "framework/contract/kungfu-contracts.registry.json",
       "artifact": "config/kungfu-contracts.registry.json",
-      "root": "sha256:1b51661a257538c04d4097132a2352b90e05645423ffb4f761deacd5a19016e5"
+      "root": "sha256:86079535834c818ba06972f3a1ff1bf8188c876746fd54d29ee04947f6de867b"
     },
     {
       "id": "work-lifecycle-operation-matrix",
