+++ scripts/check-exit-history-surfaces.test.mjs
@@ -19,8 +19,10 @@ const kfd = readJson(
 );
 const cli = read('framework/core/src/python/kungfu/cli/commands/exit.py');
 const core = read('framework/core/src/python/kungfu/exit_bundle.py');
+const history = read('framework/core/src/python/kungfu/exit_verifier.py');
 const tui = read('framework/tui/src/main.tsx');
 const gui = read('framework/gui/src/renderer/src/main.tsx');
+const guiObserver = read('framework/gui/src/runtime-status.ts');
 
 const expected = new Map([
   ['kungfu exit history status', ['read', 'kungfu.exit.bundle']],
@@ -55,24 +57,27 @@ test('CLI and KFD expose the complete history operation family', () => {
 });
 
 test('all history verbs delegate to the existing Core authority seams', () => {
-  assert.match(cli, /exit_bundle\.history_status/u);
+  assert.match(cli, /exit_verifier\.status/u);
   assert.match(cli, /exit_bundle\.build/u);
+  assert.match(cli, /exit_verifier\.record_verified_export/u);
   assert.match(cli, /exit_verifier\.verify_(?:file|bytes)/u);
   assert.match(cli, /exit_bundle\.import_package/u);
-  assert.match(cli, /exit_bundle\.rebuild_history_projections/u);
-  assert.match(core, /memberDomainsRemainAuthoritative/u);
-  assert.match(core, /observerMutation": False/u);
-  assert.match(core, /authorityMutation": False/u);
-  assert.doesNotMatch(core, /shutil\.copy.*projection/u);
+  assert.match(cli, /exit_verifier\.rebuild_projections/u);
+  assert.match(history, /memberDomainsRemainAuthoritative/u);
+  assert.match(history, /observerMutation": False/u);
+  assert.match(history, /authorityMutation": False/u);
+  assert.doesNotMatch(history, /shutil\.copy.*projection/u);
 });
 
 test('GUI and TUI observe the same status and exact next action', () => {
-  for (const source of [gui, tui]) {
-    assert.match(source, /\['exit', 'history', 'status', '--json'\]/u);
-    assert.match(source, /historyStatus\.nextActions\[0\]/u);
-    assert.match(source, /EXIT_HISTORY_STATUS_FALLBACK/u);
-    assert.match(source, /kungfu exit history status --json/u);
-  }
+  assert.match(guiObserver, /\['exit', 'history', 'status', '--json'\]/u);
+  assert.match(guiObserver, /EXIT_HISTORY_STATUS_FALLBACK/u);
+  assert.match(guiObserver, /kungfu exit history status --json/u);
+  assert.match(gui, /historyStatus\.nextActions\[0\]/u);
+  assert.match(tui, /\['exit', 'history', 'status', '--json'\]/u);
+  assert.match(tui, /historyStatus\.nextActions\[0\]/u);
+  assert.match(tui, /EXIT_HISTORY_STATUS_FALLBACK/u);
+  assert.match(tui, /kungfu exit history status --json/u);
   assert.match(tui, /History \{historyStatus\.state\}/u);
   assert.match(gui, /system\.history-protection/u);
 });
