+++ scripts/readonly-agent-bootstrap.test.mjs
@@ -234,6 +234,7 @@ test('declared discovery routes are zero-write in a cold read-only fixture', (t)
     'docs/qualification/stable-release-continuation.contract.json',
     'docs/adr/KF-ADR-019fbbe4-40c5-7c67-9ed2-910a65430ff7.md',
     'docs/adr/SHIFU-ADR-019fab1a-2853-737e-8c67-a9b1aa9035aa.md',
+    'docs/adr/KF-ADR-019fbe04-6548-7b3a-8358-22c8fe8238a9.md',
     'docs/architecture/adr-map.json',
     'docs/architecture/adr-map.md',
     'framework/release/publication-surfaces.json',
@@ -262,6 +263,8 @@ test('declared discovery routes are zero-write in a cold read-only fixture', (t)
     'crates/shifu/src/registrar.rs',
     'config/kungfu-contracts.registry.json',
     'framework/core/src/python/kungfu/distribution_update.py',
+    'framework/core/src/python/kungfu/exit_bundle.py',
+    'framework/core/src/python/kungfu/exit_verifier.py',
     'framework/core/src/python/kungfu/release_channel.py',
     'framework/core/src/python/kungfu/runtime_upgrade.py',
     'framework/core/tests/python/test_distribution_release_cut.py',
@@ -371,9 +374,12 @@ test('declared discovery routes are zero-write in a cold read-only fixture', (t)
     'docs/qualification/gates/release-admission-policy.json',
     'framework/core/src/python/kungfu/agent/documentation.py',
     'framework/core/src/python/kungfu/cli/surface_contract.py',
+    'framework/core/src/python/kungfu/cli/commands/exit.py',
     'framework/core/src/python/kungfu/cli/commands/env.py',
     'framework/core/src/python/kungfu/cli/commands/kfx.py',
     'framework/core/src/python/kungfu/cli/commands/shifu.py',
+    'framework/gui/src/renderer/src/main.tsx',
+    'framework/gui/src/runtime-status.ts',
     'framework/maintainability/baseline-transitions/README.md',
     'framework/maintainability/code-complexity-policy.json',
     'framework/maintainability/code-complexity-baseline.json',
